New notation for assigning controllers to routes.
This will allow a specific action method in a given controller to
handle a route, rather than being purely based on URL and HTTP
method.

The new format is either "controller#action" if the controller is
a class that needs to be instantiated, or "controller.action" if
the controller is a static object.

The standard "Turnpike way" would be to use a class for the
controller and thus the '#' notation.